requirements.txt
# Data extraction
yfinance>=0.2.40
requests>=2.31.0

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Orchestration (install separately if needed - has many dependencies)
# apache-airflow>=2.8.0
# apache-airflow-providers-postgres>=5.10.0

# Transformation
dbt-core>=1.7.4
dbt-postgres>=1.7.4

# Visualization
streamlit>=1.30.0
plotly>=5.18.0

# Development
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
flake8>=7.0.0
isort>=5.13.2
mypy>=1.8.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tenacity>=8.2.3
python-dateutil>=2.8.2
